moved sockets part
