+++ components/ThreeJS/ControlPanel/TabPanels/Information.tsx
@@ -8,8 +8,6 @@ import {
    currentBrowsers,
 } from '../../../../utils/recoil/browser'
 import { Text, Container } from '@mantine/core'
-import { showNotification } from '@mantine/notifications'
-import { SocketContext } from '../../../../utils/context/socket'
 import { useRecoilValue, waitForAll } from 'recoil'
 const fetchParticipantsData = async (url: string, embed_url: string) => {
    console.log('url', embed_url)
@@ -43,7 +41,7 @@ const format = (str: string) => {
    if (str.length < 15) return str
    return str.substring(0, 6) + '...'
 }
-export default function Information() {
+export default function Information({ update }) {
    const [index, browsers] = useRecoilValue(
       waitForAll([currentBrowserIndex, currentBrowsers])
    )
@@ -51,34 +49,7 @@ export default function Information() {
    const { data, mutate, isError, isLoading } = useParticipants(
       browsers[index].url
    )
-   const socket = useContext(SocketContext)
 
-   useEffect(() => {
-      socket.on('participantsAdded', (msg) => {
-         console.log('par Added', msg)
-         showNotification({
-            title: 'Member joining',
-            message: `${msg.email} is joining`,
-            color: 'blue',
-            autoClose: false,
-         })
-         mutate()
-      })
-      socket.on('participantsRemoved', (msg) => {
-         console.log('par Removed', msg)
-         showNotification({
-            title: 'Member leaving',
-            message: `${msg.email} left the session`,
-            color: 'red',
-            autoClose: false,
-         })
-         mutate()
-      })
-      return () => {
-         socket.off('participantsAdded')
-         socket.off('participantsRemoved')
-      }
-   }, [])
    return (
       <>
          <Container>
